requirements.txt
appdirs==1.4.4
argcomplete==3.5.3
backcall==0.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
docopt==0.6.2
fastjsonschema==2.20.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mercadopago==2.2.3
pickleshare==0.7.5
pillow==11.1.0
playwright==1.51.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyee==12.1.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-barcode==0.15.1
python-dotenv==1.0.1
python-escpos==3.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.38
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
